First Edit
Deleted Temporary methods

Made new method in Database: returnSuggestionID. Since createSuggestion in the Database does not return an ID, I had to find another way to obtain the ID that is generated when you create a new SuggestionData. Assuming that SuggestionData Hence, I made returnSuggestionID method to actually create the suggestion and add it to the suggestionMap. This did create two methods in the Database that creates Suggestion, so it may be problematic